Moved the scoredQuery function to db-calls instead of lib/db.js
